rename package to pushpop
