+++ src/main/java/common/pushPop/Heap.java
@@ -1,4 +1,4 @@
-package common.heap;
+package common.pushPop;
 
 import java.util.Comparator;
 import java.util.function.BiPredicate;
