+++ src/main/java/common/pushPop/HeapList.java
@@ -1,4 +1,4 @@
-package common.heap;
+package common.pushPop;
 
 import java.util.function.Supplier;
 import java.util.stream.Stream;
